requirements.txt
# Requirements for BackgroundFX Standalone Service (Gradio)

# Core App Framework
gradio==4.31.0

# Image and Video Processing
numpy==1.26.4
opencv-python-headless
Pillow
requests
tqdm

# AI Models - PyTorch and SAM2
# Railway will need a GPU plan for this to work effectively.
# We specify CUDA 12.1 versions, which are common on cloud platforms.
torch --extra-index-url https://download.pytorch.org/whl/cu121
torchvision --extra-index-url https://download.pytorch.org/whl/cu121
transformers

# SAM2 from its official source
sam2 @ git+https://github.com/facebookresearch/segment-anything-2.git
